% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capek_books.R
\name{capek_books}
\alias{capek_books}
\title{Tidy data frame of Karel Capek's 6 completed, published novels}
\usage{
capek_books()
}
\value{
A data frame with two columns: \code{text} and \code{book}
}
\description{
Returns a tidy data frame of Karel Capek's 6 completed, published novels with
two columns: \code{text}, which contains the text of the novels,
and \code{book}, which contains the titles of
the novels as a factor in order of publication.
}
\examples{

library(dplyr)

capek_books() \%>\%
     group_by(book) \%>\%
     summarise(total_lines = n())

}
